CounterButton now takes a kata as a prop, this eliminates a long of props that I was passing in. Some props remain because to remove them I need to make an object containing settings to pass in and I'm too lazy to refactor data() so that I can do that.
Updated the App so we use ID matching everywhere when editing Kata. Hopefully I didn't miss anything.

Updated the App so when we save a kata if it has no ID we will generate one with Date.now().

Updated all the tempate kata so they don't have IDs, it's better to generate them on the fly so we don't get clashes. We might get some dupes when people add kata from templates multiple times but that isn't really undesired behaviour TBH.

Updated the ReadMe, now all my features and TODOs are complete :) with the exception of making the app look extra pretty ofc, but that is something for another time